refactor(assets_generator)!: relocate module out of ai4rag.utils
Move ai4rag.utils.assets_generator to ai4rag.assets_generator so it
sits alongside other top-level packages instead of under the
general-purpose utils namespace, and update every import, docs
reference, and the setuptools package-data entry accordingly.

BREAKING CHANGE: ai4rag.utils.assets_generator no longer exists.
Import from ai4rag.assets_generator instead.
