+++ ai4rag/assets_generator/__init__.py
@@ -2,9 +2,9 @@
 # Copyright IBM Corp. 2026
 # SPDX-License-Identifier: Apache-2.0
 # -----------------------------------------------------------------------------
-from ai4rag.utils.assets_generator.leaderboard import build_leaderboard_html
-from ai4rag.utils.assets_generator.notebook import Notebook
-from ai4rag.utils.assets_generator.templates import generate_notebook_from_template
+from ai4rag.assets_generator.leaderboard import build_leaderboard_html
+from ai4rag.assets_generator.notebook import Notebook
+from ai4rag.assets_generator.templates import generate_notebook_from_template
 
 __all__ = [
     "Notebook",
