+++ docs/api-reference/assets-generator/notebooks.md
@@ -4,7 +4,7 @@ Programmatic notebook generation from RAG pattern definitions.
 
 ## Notebook Builder
 
-::: ai4rag.utils.assets_generator.notebook
+::: ai4rag.assets_generator.notebook
     options:
       members:
         - Notebook
@@ -12,7 +12,7 @@ Programmatic notebook generation from RAG pattern definitions.
 
 ## Template Rendering
 
-::: ai4rag.utils.assets_generator.templates
+::: ai4rag.assets_generator.templates
     options:
       members:
         - generate_notebook_from_template
