+++ docs/user-guide/pipeline-components.md
@@ -1,6 +1,6 @@
 # Pipeline Components
 
-`ai4rag.utils.data` and `ai4rag.utils.assets_generator` provide reusable building blocks for RAG
+`ai4rag.utils.data` and `ai4rag.assets_generator` provide reusable building blocks for RAG
 pipeline workflows. These functions encapsulate the business logic that was previously inlined in
 Kubeflow Pipeline components, making it available for use in any context — KFP pipelines, standalone
 scripts, notebooks, or tests.
@@ -23,9 +23,9 @@ scripts, notebooks, or tests.
 │  ┌─────────────┐ ┌─────────────────────────┐ │
 │  │ utils/                                    │ │
 │  │  data/                                    │ │
-│  │  assets_generator/                        │ │
-│  │   notebook, leaderboard, templates        │ │
 │  │  clients/ — s3, maas_client               │ │
+│  │ assets_generator/                         │ │
+│  │  notebook, leaderboard, templates         │ │
 │  └───────────────────────────────────────────┘ │
 │  ┌──────────────────────────────────────────┐ │
 │  │ core/ — experiment, HPO                  │ │
