+++ tests/unit/ai4rag/assets_generator/__init__.py
